+++ crates/orchestrator/src/cli/cron/mod.rs
@@ -1,0 +1,8 @@
+use crate::alerts::aws_sns::AWSSNSValidatedArgs;
+
+pub mod event_bridge;
+
+#[derive(Clone, Debug)]
+pub enum CronParams {
+    AWSSNS(AWSSNSValidatedArgs),
+}