+++ crates/orchestrator/src/queue/mod.rs
@@ -6,11 +6,125 @@ use std::time::Duration;
 
 use async_trait::async_trait;
 use color_eyre::Result as EyreResult;
+use lazy_static::lazy_static;
 use mockall::automock;
 use omniqueue::{Delivery, QueueError};
+use strum_macros::Display;
 
 use crate::config::Config;
 use crate::jobs::JobError;
+use crate::setup::SetupConfig;
+
+#[derive(Display, Debug, Clone, PartialEq, Eq)]
+pub enum QueueType {
+    #[strum(serialize = "snos_job_processing")]
+    SnosJobProcessing,
+    #[strum(serialize = "snos_job_verification")]
+    SnosJobVerification,
+    #[strum(serialize = "proving_job_processing")]
+    ProvingJobProcessing,
+    #[strum(serialize = "proving_job_verification")]
+    ProvingJobVerification,
+    #[strum(serialize = "proof_registration_job_processing")]
+    ProofRegistrationJobProcessing,
+    #[strum(serialize = "proof_registration_job_verification")]
+    ProofRegistrationJobVerification,
+    #[strum(serialize = "data_submission_job_processing")]
+    DataSubmissionJobProcessing,
+    #[strum(serialize = "data_submission_job_verification")]
+    DataSubmissionJobVerification,
+    #[strum(serialize = "update_state_job_processing")]
+    UpdateStateJobProcessing,
+    #[strum(serialize = "update_state_job_verification")]
+    UpdateStateJobVerification,
+    #[strum(serialize = "job_handle_failure")]
+    JobHandleFailure,
+    #[strum(serialize = "worker_trigger")]
+    WorkerTrigger,
+}
+
+impl QueueType {
+    pub fn iter() -> impl Iterator<Item = QueueType> {
+        [
+            QueueType::SnosJobProcessing,
+            QueueType::SnosJobVerification,
+            QueueType::ProvingJobProcessing,
+            QueueType::ProvingJobVerification,
+            QueueType::ProofRegistrationJobProcessing,
+            QueueType::ProofRegistrationJobVerification,
+            QueueType::DataSubmissionJobProcessing,
+            QueueType::DataSubmissionJobVerification,
+            QueueType::UpdateStateJobProcessing,
+            QueueType::UpdateStateJobVerification,
+            QueueType::JobHandleFailure,
+            QueueType::WorkerTrigger,
+        ]
+        .iter()
+        .cloned()
+    }
+}
+
+#[derive(Clone)]
+pub struct DlqConfig {
+    pub max_receive_count: i32,
+    pub dlq_name: QueueType,
+}
+
+#[derive(Clone)]
+pub struct QueueConfig {
+    pub name: QueueType,
+    pub visibility_timeout: i32,
+    pub dlq_config: Option<DlqConfig>,
+}
+
+// TODO: use QueueType::iter() or format!
+lazy_static! {
+    pub static ref QUEUES: Vec<QueueConfig> = vec![
+        QueueConfig {
+            name: QueueType::SnosJobProcessing,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::SnosJobVerification,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::ProvingJobProcessing,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::ProvingJobVerification,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::DataSubmissionJobProcessing,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::DataSubmissionJobVerification,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::UpdateStateJobProcessing,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig {
+            name: QueueType::UpdateStateJobVerification,
+            visibility_timeout: 300,
+            dlq_config: Some(DlqConfig { max_receive_count: 5, dlq_name: QueueType::JobHandleFailure })
+        },
+        QueueConfig { name: QueueType::JobHandleFailure, visibility_timeout: 300, dlq_config: None },
+        QueueConfig { name: QueueType::WorkerTrigger, visibility_timeout: 300, dlq_config: None },
+    ];
+}
+
 /// Queue Provider Trait
 ///
 /// The QueueProvider trait is used to define the methods that a queue
@@ -19,16 +133,17 @@ use crate::jobs::JobError;
 #[automock]
 #[async_trait]
 pub trait QueueProvider: Send + Sync {
-    async fn send_message_to_queue(
-        &self,
-        queue: job_queue::QueueType,
-        payload: String,
-        delay: Option<Duration>,
-    ) -> EyreResult<()>;
-    async fn consume_message_from_queue(
-        &self,
-        queue: job_queue::QueueType,
-    ) -> std::result::Result<Delivery, QueueError>;
+    async fn send_message_to_queue(&self, queue: QueueType, payload: String, delay: Option<Duration>)
+    -> EyreResult<()>;
+    async fn consume_message_from_queue(&self, queue: QueueType) -> std::result::Result<Delivery, QueueError>;
+    async fn create_queue(&self, queue_config: &QueueConfig, config: &SetupConfig) -> EyreResult<()>;
+    async fn setup(&self, config: SetupConfig) -> EyreResult<()> {
+        // Creating the queues :
+        for queue in QUEUES.iter() {
+            self.create_queue(queue, &config).await?;
+        }
+        Ok(())
+    }
 }
 
 pub async fn init_consumers(config: Arc<Config>) -> Result<(), JobError> {