+++ crates/orchestrator/src/tests/workers/proving/mod.rs
@@ -25,7 +25,7 @@ use crate::workers::proving::ProvingWorker;
 #[case(false)]
 #[tokio::test]
 async fn test_proving_worker(#[case] incomplete_runs: bool) -> Result<(), Box<dyn Error>> {
-    use crate::queue::job_queue::QueueType;
+    use crate::queue::QueueType;
 
     let server = MockServer::start();
     let da_client = MockDaClient::new();